Moved Plink class to tools.py
